Moved HQL logging from an anonymous function to a Java class so it can be referenced in logback.xml
