+++ elide-datastore/elide-datastore-hibernate/src/main/java/com/yahoo/elide/core/hibernate/QueryLogger.java
@@ -3,7 +3,7 @@
  * Licensed under the Apache License, Version 2.0
  * See LICENSE file in project root for terms.
  */
-package com.yahoo.elide.datastores.jpa.porting;
+package com.yahoo.elide.core.hibernate;
 
 /**
  * Logs JPQL queries from the JPA Store.
