+++ elide-datastore/elide-datastore-jpa/src/main/java/com/yahoo/elide/datastores/jpa/porting/EntityManagerWrapper.java
@@ -7,6 +7,7 @@ package com.yahoo.elide.datastores.jpa.porting;
 
 import static com.yahoo.elide.datastores.jpa.JpaDataStore.DEFAULT_LOGGER;
 import com.yahoo.elide.core.hibernate.Query;
+import com.yahoo.elide.core.hibernate.QueryLogger;
 import com.yahoo.elide.core.hibernate.Session;
 import lombok.extern.slf4j.Slf4j;
