+++ elide-datastore/elide-datastore-jpa/src/main/java/com/yahoo/elide/datastores/jpa/porting/QueryWrapper.java
@@ -7,7 +7,6 @@ package com.yahoo.elide.datastores.jpa.porting;
 
 import com.yahoo.elide.core.hibernate.Query;
 import com.yahoo.elide.datastores.jpa.ScrollableIterator;
-
 import lombok.Getter;
 
 import java.util.Collection;
